pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "identark-sdk"
version = "1.1.0"
description = "The AgentGateway Protocol — secure, scalable AI agent execution infrastructure"
readme = "README.md"
license = { text = "AGPL-3.0-only" }
authors = [{ name = "Gold Okpa", email = "gold@identark.io" }]
keywords = [
    "ai", "agents", "llm", "sandbox", "infrastructure",
    "openai", "anthropic", "mistral", "langchain", "gateway",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.26.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.18.0"]
mistral = ["mistralai>=1.0.0"]
gemini = ["google-generativeai>=0.5.0"]
all = ["openai>=1.0.0", "anthropic>=0.18.0", "mistralai>=1.0.0", "google-generativeai>=0.5.0", "langgraph>=0.2.0", "langchain-core>=0.2.0", "llama-index-core>=0.10.0"]
langchain = [
    "langchain-core>=0.2.0",
]
llamaindex = [
    "llama-index-core>=0.10.0",
]
langgraph = [
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
]
dev = [
    "identark-sdk[all]",
    "langchain-core>=0.2.0",
    "llama-index-core>=0.10.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.9",
    "ruff>=0.4",
    "pre-commit>=3.7",
    "aiofiles>=23.0",
]

[project.urls]
Homepage      = "https://github.com/identark/sdk"
Documentation = "https://github.com/identark/sdk#readme"
Repository    = "https://github.com/identark/sdk"
Issues        = "https://github.com/identark/sdk/issues"
Changelog     = "https://github.com/identark/sdk/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["identark"]

# ── Ruff (linting + formatting) ───────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "ANN"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S"]

# ── Mypy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version    = "3.10"
strict            = true
ignore_missing_imports = true
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = "identark.integrations.*"
disable_error_code = ["override", "arg-type", "misc", "assignment"]

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode    = "auto"
testpaths       = ["tests"]
addopts         = "-v --tb=short"
filterwarnings  = ["ignore::DeprecationWarning"]
markers = [
    "integration: integration test requiring mock server",
]

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["identark"]
omit   = ["identark/integrations/*", "tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "\\.\\.\\.\\s*$",
]
